Create Overlap Finder
This function can find overlaps between the beginning and the end of given DNA sequences from a FASTA format. 

Variables required are the file name which sequences are located, and the overlap length which you would like to find. 

This function was created as a solution to http://rosalind.info/problems/grph/